requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
geographiclib==2.0
geopy==2.4.1
h3==4.3.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.41
sqlalchemy2-stubs==0.0.2a38
timezonefinder==6.6.3
typing_extensions==4.14.1
Werkzeug==3.1.3
zipp==3.23.0
gunicorn